requirements.txt
requests
beautifulsoup4
transformers
alpaca-trade-api
python-dotenv